Utility effect implementation part 2

+++ Quest-for-eternity/Assets/DamageBoostEffect.cs
@@ -4,12 +4,12 @@ using UnityEngine;
 [System.Serializable]
 public class DamageBoostEffect : EffectTemplate
 {
-    public override void UseEffect<T>(int targetId, int effectValue, T value)
+    public override void UseEffect<T>(int targetId, float effectValue, T value)
     {
         Debug.Log("damage boost activated");
        // GameObject inputGameobject = value as GameObject;
 
-        FieldScript.damagePoints += effectValue;
+        FieldScript.damagePoints += (int)effectValue;
         /*if (inputGameobject.GetComponent<SceneObjectDatabase>() != null)
         {
