+++ Quest-for-eternity/Assets/EffectTemplate.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public abstract class EffectTemplate : ScriptableObject
 {
     
-    public virtual void UseEffect<T>(int targetId,int effectValue ,T value)
+    public virtual void UseEffect<T>(int targetId,float effectValue ,T value)
     {
         Debug.Log("I dun it!");
     }
